// src/app/models/mortality-data.ts
export interface MortalityData {
    age: number;
    sex: number;
    bmi: number;
    hypertensive: number;
    atrial_fbrillation: number;
    diabetes: number;
    heartbeat: number;
    systole: number;
    diastole: number;
    breathing_rate: number;
    temperature: number;
    oxygen_saturation: number;
    leucocyte: number;
    urea_nitrogen: number;
    blood_glucose: number;
    anion_gap: number;
    bicarbonate: number;
    lactic_acid: number;
    mortality_risk: number;
    health_data: number;
}